Remove Nunjucks
